pyproject.toml
[tool.poetry]
name = "algorithms"
version = "0.1.0"
description = ""
authors = ["xavi <xpeiro92@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.1.2"
pydantic = "^2.11"
PyHamcrest = "^2.1.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
